permissions/revocation.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Permissions API - Revocation Tests</title>
<link rel="help" href="https://www.w3.org/TR/permissions/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>
<script>
'use strict';

// 1. "granted" -> "prompt"
promise_test(async t => {
  await test_driver.set_permission({ name: 'geolocation' }, 'granted');
  const status = await navigator.permissions.query({ name: 'geolocation' });
  assert_equals(status.state, 'granted', 'Initial state is "granted"');

  const p = new Promise(resolve => {
    const listener = () => {
      status.removeEventListener('change', listener);
      resolve();
    };
    status.addEventListener('change', listener);
  });

  await test_driver.set_permission({ name: 'geolocation' }, 'prompt');
  await p;

  assert_equals(status.state, 'prompt', 'Now in "prompt" state');
  
  // Cleanup
  await test_driver.set_permission({ name: 'geolocation' }, 'prompt');
}, 'Transition "granted" -> "prompt" fires a "change" event');

// 2. "granted" -> "denied"
promise_test(async t => {
  await test_driver.set_permission({ name: 'geolocation' }, 'granted');
  const status = await navigator.permissions.query({ name: 'geolocation' });
  assert_equals(status.state, 'granted', 'Initial state is "granted"');

  const p = new Promise(resolve => {
    const listener = () => {
      status.removeEventListener('change', listener);
      resolve();
    };
    status.addEventListener('change', listener);
  });

  await test_driver.set_permission({ name: 'geolocation' }, 'denied');
  await p;

  assert_equals(status.state, 'denied', 'Now in "denied" state');

  // Cleanup
  await test_driver.set_permission({ name: 'geolocation' }, 'prompt');
}, 'Transition "granted" -> "denied" fires a "change" event');
</script>
</body>
</html>
